+++ Craftsman/Builders/MassTransitModifier.cs
@@ -70,5 +70,61 @@
                 File.Move(tempPath, classPath.FullClassPath);
             }
         }
+
+        public static void AddProducerRegistation(string solutionDirectory, string endpointRegistrationName, string projectBaseName)
+        {
+            var classPath = ClassPathHelper.WebApiServiceExtensionsClassPath(solutionDirectory, $"{Utilities.GetMassTransitRegistrationName()}.cs", projectBaseName);
+
+            if (!Directory.Exists(classPath.ClassDirectory))
+                throw new DirectoryNotFoundException($"The `{classPath.ClassDirectory}` directory could not be found.");
+
+            if (!File.Exists(classPath.FullClassPath))
+                throw new FileNotFoundException($"The `{classPath.FullClassPath}` file could not be found.");
+
+            var producerClassPath = ClassPathHelper.WebApiProducersServiceExtensionsClassPath(solutionDirectory, $"{endpointRegistrationName}.cs", projectBaseName);
+
+            var tempPath = $"{classPath.FullClassPath}temp";
+            var hasUsingForProducerNamespace = false;
+            using (var input = File.OpenText(classPath.FullClassPath))
+            {
+                using var output = new StreamWriter(tempPath);
+                string line;
+                while (null != (line = input.ReadLine()))
+                {
+                    var newText = $"{line}";
+                    if (line.Contains($"// Producers -- Do Not Delete This Comment"))
+                        newText += @$"{Environment.NewLine}                    cfg.{endpointRegistrationName}();";
+                    if (line.Contains(producerClassPath.ClassNamespace))
+                        hasUsingForProducerNamespace = true;
+
+                    output.WriteLine(newText);
+                }
+            }
+
+            // delete the old file and set the name of the new one to the original name
+            File.Delete(classPath.FullClassPath);
+            File.Move(tempPath, classPath.FullClassPath);
+
+            if (!hasUsingForProducerNamespace)
+            {
+                using (var input = File.OpenText(classPath.FullClassPath))
+                {
+                    using var output = new StreamWriter(tempPath);
+                    string line;
+                    while (null != (line = input.ReadLine()))
+                    {
+                        var newText = $"{line}";
+                        if (line.Contains($"using MassTransit;"))
+                            newText += @$"{Environment.NewLine}    using {producerClassPath.ClassNamespace};";
+
+                        output.WriteLine(newText);
+                    }
+                }
+
+                // delete the old file and set the name of the new one to the original name
+                File.Delete(classPath.FullClassPath);
+                File.Move(tempPath, classPath.FullClassPath);
+            }
+        }
     }
 }