+++ CHANGELOG.md
@@ -32,10 +32,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Moved Service Registrations to separate files
 - Fixed default value for strings on entities to use quotes
 - Added `add:bus` command
-- Added Bus, Messages, and Consumers props to BC template
+- Added Bus, Messages, Producers, and Consumers props to BC template
 - Added `add:message` command
 - Updated entity name and entity prop names first letter to always be capitalized
-- Added `register:consumer` command
+- Added `register:consumer` and `register:producer` commands
 
 ## [0.9.3] - 2021-04-10